extract method
