+++ Ve.Metrics.StatsDClient.CastleWindsor/BaseInterceptor.cs
@@ -1,4 +1,5 @@
-﻿using System.Linq;
+﻿using System.Collections.Generic;
+using System.Linq;
 using Castle.DynamicProxy;
 using Ve.Metrics.StatsDClient.Abstract;
 using Ve.Metrics.StatsDClient.Abstract.Attributes;
@@ -36,6 +37,22 @@ namespace Ve.Metrics.StatsDClient.CastleWindsor
                     .GetMethod(methodName)
                     .GetCustomAttributes(typeof(T), true)
                     .FirstOrDefault() as T;
+        }
+
+        protected static string GetGenericName(string name, System.Reflection.MethodBase methodBase)
+        {
+            var generic = new List<string>();
+            var arguments = methodBase.GetGenericArguments();
+
+            while (arguments.Any())
+            {
+                var argument = arguments.First();
+                generic.Add(argument.Name.ToLowerInvariant());
+                arguments = argument.GetGenericArguments();
+            }
+
+            name = name.Replace("{generic}", string.Join("-", generic));
+            return name;
         }
     }
 }
