+++ Ve.Metrics.StatsDClient.CastleWindsor/StatsDTimingInterceptor.cs
@@ -1,6 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
+﻿using System.Diagnostics;
 using Castle.DynamicProxy;
 using Ve.Metrics.StatsDClient.Abstract;
 using Ve.Metrics.StatsDClient.Abstract.Attributes;
@@ -30,17 +28,7 @@ namespace Ve.Metrics.StatsDClient.CastleWindsor
 
             if (methodBase.IsGenericMethod)
             {
-                var generic = new List<string>();
-                var arguments = methodBase.GetGenericArguments();
-
-                while (arguments.Any())
-                {
-                    var argument = arguments.First();
-                    generic.Add(argument.Name.ToLowerInvariant());
-                    arguments = argument.GetGenericArguments();
-                }
-
-                name = name.Replace("{generic}", string.Join("-", generic));
+                name = GetGenericName(name, methodBase);
             }
 
             Client.LogTiming(name, watch.ElapsedMilliseconds, attr.Tags);
